+++ src/connection/etl/import/options.rs
@@ -5,7 +5,11 @@ use arrayvec::ArrayString;
 use super::{ExaImport, Trim};
 use crate::{
     connection::etl::RowSeparator,
-    etl::{job::EtlJob, JobFuture, WithSocketFuture},
+    etl::{
+        import::ExaImportState,
+        job::{EtlJob, SocketSetup},
+        EtlQuery,
+    },
     ExaConnection, SqlxResult,
 };
 
@@ -61,11 +65,11 @@ impl<'a> ImportBuilder<'a> {
     pub async fn build<'c>(
         &'a self,
         con: &'c mut ExaConnection,
-    ) -> SqlxResult<(JobFuture<'c>, Vec<ExaImport>)>
+    ) -> SqlxResult<(EtlQuery<'c>, Vec<ExaImport>)>
     where
         'c: 'a,
     {
-        self.build_etl(con).await
+        self.build_job(con).await
     }
 
     /// Sets the number of writer jobs that will be started.
@@ -147,8 +151,12 @@ impl EtlJob for ImportBuilder<'_> {
         self.num_writers
     }
 
-    fn create_worker(&self, future: WithSocketFuture, with_compression: bool) -> Self::Worker {
-        ExaImport::Setup(future, self.buffer_size, with_compression)
+    fn create_worker(&self, setup_future: SocketSetup, with_compression: bool) -> Self::Worker {
+        ExaImport(ExaImportState::Setup(
+            setup_future,
+            self.buffer_size,
+            with_compression,
+        ))
     }
 
     fn query(&self, addrs: Vec<SocketAddrV4>, with_tls: bool, with_compression: bool) -> String {